Move plugin discovery behavior tests to core-plugins
