+++ codex-rs/core-plugins/Cargo.toml
@@ -51,4 +51,6 @@ zip = { workspace = true }
 libc = { workspace = true }
 pretty_assertions = { workspace = true }
 tempfile = { workspace = true }
+tracing-subscriber = { workspace = true }
+tracing-test = { workspace = true, features = ["no-env-filter"] }
 wiremock = { workspace = true }
